Move RaftStoreRouter to raftstore
