+++ cmd/src/bin/tikv-ctl.rs
@@ -36,9 +36,10 @@ use kvproto::tikvpb::TikvClient;
 use pd_client::{Config as PdConfig, PdClient, RpcClient};
 use raft::eraftpb::{ConfChange, Entry, EntryType};
 use tikv::config::TiKvConfig;
+use tikv::raftstore::router::ServerRaftStoreRouter;
 use tikv::raftstore::store::{keys, INIT_EPOCH_CONF_VER};
 use tikv::server::debug::{BottommostLevelCompaction, Debugger, RegionInfo};
-use tikv::server::{RaftKv, ServerRaftStoreRouter};
+use tikv::server::RaftKv;
 use tikv::storage::kv::Engine;
 use tikv_util::security::{SecurityConfig, SecurityManager};
 use tikv_util::{escape, unescape};
