+++ cmd/src/server.rs
@@ -24,6 +24,7 @@ use tikv::config::TiKvConfig;
 use tikv::coprocessor;
 use tikv::import::{ImportSSTService, SSTImporter};
 use tikv::raftstore::coprocessor::{CoprocessorHost, RegionInfoAccessor};
+use tikv::raftstore::router::ServerRaftStoreRouter;
 use tikv::raftstore::store::fsm::store::{StoreMeta, PENDING_VOTES_CAP};
 use tikv::raftstore::store::{fsm, LocalReader};
 use tikv::raftstore::store::{new_compaction_listener, SnapManagerBuilder};
@@ -32,7 +33,6 @@ use tikv::server::lock_manager::LockManager;
 use tikv::server::resolve;
 use tikv::server::service::{DebugService, DiagnosticsService};
 use tikv::server::status_server::StatusServer;
-use tikv::server::transport::ServerRaftStoreRouter;
 use tikv::server::DEFAULT_CLUSTER_ID;
 use tikv::server::{create_raft_storage, Node, RaftKv, Server};
 use tikv::storage;
