+++ src/raftstore/mod.rs
@@ -2,6 +2,7 @@
 
 pub mod coprocessor;
 pub mod errors;
+pub mod router;
 pub mod store;
 pub use self::coprocessor::{RegionInfo, RegionInfoAccessor, SeekRegionCallback};
 pub use self::errors::{DiscardReason, Error, Result};
