+++ src/server/gc_worker.rs
@@ -21,10 +21,10 @@ use kvproto::metapb;
 use log_wrappers::DisplayValue;
 use raft::StateRole;
 
+use crate::raftstore::router::ServerRaftStoreRouter;
 use crate::raftstore::store::keys;
 use crate::raftstore::store::msg::StoreMsg;
 use crate::raftstore::store::util::find_peer;
-use crate::server::transport::ServerRaftStoreRouter;
 use crate::storage::kv::{
     Engine, Error as EngineError, ErrorInner as EngineErrorInner, RegionInfoProvider, ScanMode,
     Statistics,
