+++ src/server/raft_client.rs
@@ -9,7 +9,7 @@ use std::time::Instant;
 use super::load_statistics::ThreadLoad;
 use super::metrics::*;
 use super::{Config, Result};
-use crate::server::transport::RaftStoreRouter;
+use crate::raftstore::router::RaftStoreRouter;
 use crossbeam::channel::SendError;
 use futures::{future, stream, Future, Poll, Sink, Stream};
 use grpcio::{
