+++ tests/benches/misc/raftkv/mod.rs
@@ -17,13 +17,13 @@ use engine::{ALL_CFS, CF_DEFAULT};
 use engine_rocks::RocksSnapshot;
 use engine_traits::Snapshot;
 use keys::Key;
+use tikv::raftstore::router::RaftStoreRouter;
 use tikv::raftstore::store::{
     cmd_resp, util, Callback, CasualMessage, RaftCommand, ReadResponse, RegionSnapshot,
     SignificantMsg, WriteResponse,
 };
 use tikv::raftstore::Result;
 use tikv::server::raftkv::{CmdRes, RaftKv};
-use tikv::server::transport::RaftStoreRouter;
 use tikv::storage::kv::{Callback as EngineCallback, CbContext, Modify, Result as EngineResult};
 use tikv::storage::Engine;
